requirements.txt
Django==1.10.5
esptool==1.3
paho-mqtt==1.2
pyserial==3.2.1
dj-database-url==0.4.1
gunicorn==19.6.0
psycopg2==2.6.1
whitenoise==2.0.6